+++ src/main/java/gregtech/api/metatileentity/BaseMetaPipeEntity.java
@@ -220,10 +220,10 @@ public class BaseMetaPipeEntity extends BaseTileEntity implements IGregTechTileE
                         oX = xCoord;
                         oY = yCoord;
                         oZ = zCoord;
-                        if (isServerSide()) for (byte i = 0; i < 6; i++)
-                            if (getCoverIDAtSide(i) != 0)
-                                if (!mMetaTileEntity.allowCoverOnSide(i, new GT_ItemStack(getCoverIDAtSide(i))))
-                                    dropCover(i, i, true);
+                        if (isServerSide()) for (int i = 0; i < 6; i++)
+                            if (getCoverIDAtSide((byte)i) != 0)
+                                if (!mMetaTileEntity.allowCoverOnSide((byte)i, new GT_ItemStack(getCoverIDAtSide((byte)i))))
+                                    dropCover((byte)i, (byte)i, true);
                         worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
                         mMetaTileEntity.onFirstTick(this);
                         if (!hasValidMetaTileEntity()) return;
@@ -249,13 +249,13 @@ public class BaseMetaPipeEntity extends BaseTileEntity implements IGregTechTileE
                     case 6:
                     case 7:
                         if (isServerSide() && mTickTimer > 10) {
-                            for (byte i = (byte) (tCode - 2); i < 6; i++)
-                                if (getCoverIDAtSide(i) != 0) {
+                            for (int i = (byte) (tCode - 2); i < 6; i++)
+                                if (getCoverIDAtSide((byte)i) != 0) {
                                     tCode++;
-                                    GT_CoverBehavior tCover = getCoverBehaviorAtSide(i);
-                                    int tCoverTickRate = tCover.getTickRate(i, getCoverIDAtSide(i), mCoverData[i], this);
+                                    GT_CoverBehavior tCover = getCoverBehaviorAtSide((byte)i);
+                                    int tCoverTickRate = tCover.getTickRate((byte)i, getCoverIDAtSide((byte)i), mCoverData[i], this);
                                     if (tCoverTickRate > 0 && mTickTimer % tCoverTickRate == 0) {
-                                        mCoverData[i] = tCover.doCoverThings(i, getInputRedstoneSignal(i), getCoverIDAtSide(i), mCoverData[i], this, mTickTimer);
+                                        mCoverData[i] = tCover.doCoverThings((byte)i, getInputRedstoneSignal((byte)i), getCoverIDAtSide((byte)i), mCoverData[i], this, mTickTimer);
                                         if (!hasValidMetaTileEntity()) return;
                                     }
                                 }
